Order total captured performance (#7166)
* Benchmarks for totalCaptured, totalAuthorized, events and subtotal on Order type

* Performance improvement for order totalCaptured resolver

* Move fixme placing

* Move model helpers out of utils, move PaymentsByOrderIdLoader to payments
