+++ saleor/graphql/order/types.py
@@ -22,6 +22,8 @@ from ...order import OrderStatus, models
 from ...order.models import FulfillmentStatus
 from ...order.utils import get_order_country, get_valid_shipping_methods_for_order
 from ...payment import ChargeStatus
+from ...payment.dataloaders import PaymentsByOrderIdLoader
+from ...payment.model_helpers import get_total_captured
 from ...product.product_images import get_product_image_thumbnail
 from ..account.dataloaders import AddressByIdLoader, UserByUserIdLoader
 from ..account.types import User
@@ -56,7 +58,6 @@ from .dataloaders import (
     OrderEventsByOrderIdLoader,
     OrderLineByIdLoader,
     OrderLinesByOrderIdLoader,
-    PaymentsByOrderIdLoader,
 )
 from .enums import OrderEventsEmailsEnum, OrderEventsEnum
 from .utils import validate_draft_order
@@ -806,9 +807,15 @@ class Order(CountableDjangoObjectType):
         return root.total_authorized
 
     @staticmethod
-    def resolve_total_captured(root: models.Order, _info):
-        # FIXME adjust to multiple payments in the future
-        return root.total_captured
+    def resolve_total_captured(root: models.Order, info):
+        def _resolve_total_captured(payments):
+            return get_total_captured(payments, root.currency)
+
+        return (
+            PaymentsByOrderIdLoader(info.context)
+            .load(root.id)
+            .then(_resolve_total_captured)
+        )
 
     @staticmethod
     def resolve_total_balance(root: models.Order, _info):
